+++ Command/CreateDatabaseDoctrineCommand.php
@@ -34,19 +34,17 @@ class CreateDatabaseDoctrineCommand extends DoctrineCommand
     {
         $this
             ->setName('doctrine:database:create')
-            ->setDescription('Creates the configured databases')
+            ->setDescription('Creates the configured database')
             ->addOption('connection', null, InputOption::VALUE_OPTIONAL, 'The connection to use for this command')
             ->addOption('if-not-exists', null, InputOption::VALUE_NONE, 'Don\'t trigger an error, when the database already exists')
             ->setHelp(<<<EOT
-The <info>doctrine:database:create</info> command creates the default
-connections database:
+The <info>%command.name%</info> command creates the default connections database:
 
-<info>php app/console doctrine:database:create</info>
+    <info>php %command.full_name%</info>
 
-You can also optionally specify the name of a connection to create the
-database for:
+You can also optionally specify the name of a connection to create the database for:
 
-<info>php app/console doctrine:database:create --connection=default</info>
+    <info>php %command.full_name% --connection=default</info>
 EOT
         );
     }
@@ -56,7 +54,12 @@ EOT
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $connection = $this->getDoctrineConnection($input->getOption('connection'));
+        $connectionName = $input->getOption('connection');
+        if (empty($connectionName) === true) {
+            $connectionName = $this->getContainer()->get('doctrine')->getDefaultConnectionName();
+        }
+        $connection = $this->getDoctrineConnection($connectionName);
+
         $ifNotExists = $input->getOption('if-not-exists');
 
         $params = $connection->getParams();
@@ -71,6 +74,7 @@ EOT
         unset($params['dbname']);
 
         $tmpConnection = DriverManager::getConnection($params);
+        $shouldNotCreateDatabase = $ifNotExists && in_array($name, $tmpConnection->getSchemaManager()->listDatabases());
 
         // Only quote if we don't have a path
         if (!isset($params['path'])) {
@@ -79,16 +83,14 @@ EOT
 
         $error = false;
         try {
-            $shouldNotCreateDatabase = $ifNotExists && in_array($name, $tmpConnection->getSchemaManager()->listDatabases());
-
             if ($shouldNotCreateDatabase) {
-                $output->writeln(sprintf('<info>Database for connection named <comment>%s</comment> already exists. Skipped.</info>'));
+                $output->writeln(sprintf('<info>Database <comment>%s</comment> for connection named <comment>%s</comment> already exists. Skipped.</info>', $name, $connectionName));
             } else {
                 $tmpConnection->getSchemaManager()->createDatabase($name);
-                $output->writeln(sprintf('<info>Created database for connection named <comment>%s</comment></info>', $name));
+                $output->writeln(sprintf('<info>Created database <comment>%s</comment> for connection named <comment>%s</comment></info>', $name, $connectionName));
             }
         } catch (\Exception $e) {
-            $output->writeln(sprintf('<error>Could not create database for connection named <comment>%s</comment></error>', $name));
+            $output->writeln(sprintf('<error>Could not create database <comment>%s</comment> for connection named <comment>%s</comment></error>', $name, $connectionName));
             $output->writeln(sprintf('<error>%s</error>', $e->getMessage()));
             $error = true;
         }